Added option to play either side, and option to play again
The user can pick a side to play at the start, and can play again without re-opening the executable. After testing, the bot can play perfectly as both crosses or circles.
I also added "-static-libgcc -static-libstdc++" to the makefile to prevent an error when I tried to run the executable files, not sure if this can cause other problems but at least the program runs.